refactor: extract shared AsyncImageLoader
PreviewView and FullscreenViewer had near-duplicate CGImageSource
thumbnail loaders differing only in max pixel size (2000 vs 4000) and
whether to prefer the embedded preview. Extract a single
AsyncImageLoader namespace parameterized by a Size preset so the
decode pipeline lives in one place.
